Move queryLoopData related functions
